ml_model/requirements.txt
tensorflow==2.9.3
nltk==3.5
keras==2.4.3
numpy==1.19.5
streamlit==0.52.1
seaborn==0.11.1
tweepy==3.10.0
textblob==0.15.3
flask==1.1.2
pandas==1.2.2
matplotlib==3.5.1
scikit-learn==0.24.1
statsmodels==0.12.2
yfinance==0.1.63
alpha-vantage==2.3.1
tweet-preprocessor @ https://pypi.anaconda.org/berber/simple/tweet-preprocessor/0.5.0/tweet-preprocessor-0.5.0.tar.gz
